# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/users:
    get:
      operationId: listCompanyUsers
      tags: [users]
      summary: List users linked to a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of company users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyUser"
    post:
      operationId: addCompanyUser
      tags: [users]
      summary: Add or invite a user to a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
  /companies/{companyId}/users/{userId}/reset-password:
    post:
      operationId: resetCompanyUserPassword
      tags: [users]
      summary: Reset a company user's password (super admin only)
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /companies/{companyId}/users/{userId}:
    patch:
      operationId: updateCompanyUser
      tags: [users]
      summary: Update user role or status in a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
    delete:
      operationId: removeCompanyUser
      tags: [users]
      summary: Remove user from a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Structures ─────────────────────────────────────────────────────────────
  /companies/{companyId}/structures:
    get:
      operationId: listStructures
      tags: [structures]
      summary: List bookable structures for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of structures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Structure"
    post:
      operationId: createStructure
      tags: [structures]
      summary: Create a bookable structure
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StructureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Structure"
  /companies/{companyId}/structures/{structureId}:
    get:
      operationId: getStructure
      tags: [structures]
      summary: Get a structure
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: structureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Structure"
    patch:
      operationId: updateStructure
      tags: [structures]
      summary: Update a structure
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: structureId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StructureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Structure"
    delete:
      operationId: deleteStructure
      tags: [structures]
      summary: Remove a structure
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: structureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reservations ───────────────────────────────────────────────────────────
  /companies/{companyId}/reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          description: Conflict - time slot already booked
  /companies/{companyId}/reservations/{reservationId}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation details
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /companies/{companyId}/reservations/{reservationId}/status:
    patch:
      operationId: updateReservationStatus
      tags: [reservations]
      summary: Confirm, cancel or reject a reservation
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listAdminUsers
      tags: [users]
      summary: List all super admin users
      responses:
        "200":
          description: Super admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [users]
      summary: Create a new super admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "409":
          description: Phone already in use
  /users/{userId}:
    delete:
      operationId: deleteAdminUser
      tags: [users]
      summary: Delete a super admin user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Plans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
    AdminUser:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
    CreateAdminUserInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
...
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 6
          description: "New password to set for the user"
...
    CompanyUser:
      type: object
      required: [id, userId, companyId, role, status, name, phone]
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          description: "admin, operator, member"
        status:
          type: string
          description: "active, inactive"
        createdAt:
          type: string
          format: date-time
    CompanyUserInput:
      type: object
      required: [name, phone, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          description: "admin, operator, member"
    CompanyUserUpdate:
      type: object
      properties:
        role:
          type: string
        status:
          type: string

    # ── Structures ────────────────────────────────────────────────────────────
    Structure:
      type: object
      required: [id, companyId, name, type, status, capacity]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "court, pool, room, field, gym, other"
        description:
          type: ["string", "null"]
        capacity:
          type: integer
        status:
          type: string
          description: "active, inactive, maintenance"
        openTime:
          type: ["string", "null"]
          description: "HH:MM format"
        closeTime:
          type: ["string", "null"]
          description: "HH:MM format"
        maxDurationMinutes:
          type: ["integer", "null"]
        requiresApproval:
          type: boolean
        createdAt:
          type: string
          format: date-time
    StructureInput:
      type: object
      required: [name, type, capacity]
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        capacity:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
        maxDurationMinutes:
          type: integer
        requiresApproval:
          type: boolean
    StructureUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
        status:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        maxDurationMinutes:
          type: integer
        requiresApproval:
          type: boolean

    # ── Reservations ──────────────────────────────────────────────────────────
    Reservation:
      type: object
      required: [id, companyId, structureId, userId, date, startTime, endTime, status, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        structureId:
          type: integer
        structureName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          description: "HH:MM"
        endTime:
          type: string
          description: "HH:MM"
        status:
          type: string
          description: "requested, confirmed, cancelled, rejected, completed"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReservationInput:
      type: object
      required: [structureId, date, startTime, endTime]
      properties:
        structureId:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
    ReservationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: "confirmed, cancelled, rejected, completed"
        reason:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...