# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  structureId: number;
  date: Date;
  startTime: string;
  endTime: string;
  notes?: string;
}